refactor(launcher): extract SettingsPageMixin; main_app 3854 -> 2409 lines
Moves the whole settings cluster (_page_settings + its 23 helpers: index /
device / accel / perf-preset / update-check / monitor) verbatim into
launcher/pages/settings_page.py. main_app.py is now under half its original
size (4461 -> 2409). Behaviour-preserving: methods resolve via self on the
composed instance.

Verification upgraded from import-only to pyflakes static analysis, which
catches undefined names used inside method bodies (the runtime-NameError
risk a bare import check misses). The mixin's imports were reconciled to
zero F821/F401; the only remaining warnings are 8 pre-existing
e-in-lambda cases confirmed present in HEAD's main_app.py before the move.
python3.12 import + MRO green, full pytest suite green (120).
